+++ corpus/analyze-method-call/.clj-kondo/checking.clj
@@ -1,0 +1,5 @@
+(ns checking)
+
+(defn checking
+  [_]
+  (throw (ex-info "Use deterministic version" {})))