+++ corpus/analyze-method-call/src/test.clj
@@ -1,0 +1,4 @@
+(ns test)
+
+(System/currentTimeMillis)
+(java.util.UUID/randomUUID)