+++ CHANGELOG.md
@@ -19,6 +19,9 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 ## Unreleased
 
+- [#916](https://github.com/clj-kondo/clj-kondo/issues/916): New linter: `:destructured-or-binding-of-same-map` which warns about
+  `:or` defaults referring to bindings of same map, which is undefined and may result in broken
+  behavior
 - [#2362](https://github.com/clj-kondo/clj-kondo/issues/2362): turn min-version warning into lint warning
 
 ## 2024.08.01