+++ CHANGELOG.md
@@ -25,6 +25,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - [#2362](https://github.com/clj-kondo/clj-kondo/issues/2362): turn min-version warning into lint warning
 - [#1603](https://github.com/clj-kondo/clj-kondo/issues/1603): Support Java classes in `:analyze-call` hook
 - [#2369](https://github.com/clj-kondo/clj-kondo/issues/2369): false positive unused value in quoted list
+- [#2374](https://github.com/clj-kondo/clj-kondo/issues/2374): Detect suspicious return Schemas ([@frenchy64](https://github.com/frenchy64))
 
 ## 2024.08.01